requirements.txt
mitmproxy~=10.1.6
pycryptodome>=3.14.0
requests>=2.27.1
colorama~=0.4.6
aiohttp>=3.8.1
msgpack~=1.0.7